neural-network/requirements.txt
tensorflow-macos>=2.15.0
tensorflow-metal>=1.1.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0 
seaborn>=0.13.0
python-binance==1.0.19